:art: Move timefilter func to functions
